requirements.txt
aiohttp>=3.9
aiosqlite>=0.19
beautifulsoup4>=4.12
playwright>=1.40
pyyaml>=6.0
pydantic>=2.0
